third_party/WebKit/LayoutTests/platform/win/css3/motion-path/offset-rotation-expected.txt

CONSOLE WARNING: offset-rotation is deprecated and will be removed in M58, around April 2017. Please use offset-rotate instead. See https://www.chromestatus.com/features/6390764217040896 for more details.
This is a testharness.js-based test.
PASS offset-rotation default is auto 
PASS offset-rotation auto expands to auto 0deg 
PASS offset-rotation reverse expands to auto 180deg 
PASS offset-rotation can be a fixed angle 
PASS offset-rotation angles are converted to degrees 
PASS offset-rotation can be supplied with angle before auto 
PASS offset-rotation can be supplied with angle before reverse 
PASS offset-rotation is unaffected by child elements 
PASS offset-rotation can be explicitly inherited 
PASS offset-rotation is not inherited by default 
PASS offset-rotation style can be set inline 
Harness: the test ran to completion.
